requirements.txt
pymongo>=4.0.0
mysql-connector-python>=8.0.0
psycopg2-binary>=2.9.0
pydantic>=2.0.0
inflect>=5.0.0
python-dotenv>=0.19.0 